pyproject.toml
[project]
name = "iot-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-zerobus-ingest-sdk>=0.2.0",
    "python-dotenv>=1.2.1",
]
